Refactor: add method to select person, create person controller, fix typo in the rantals class.
